Add third/oper-tag, third/hiddenoper-tag, and third/id-tag modules
* third/oper-tag implements the solanum.chat/oper capability and
message tag, which is added to all messages sent by non-hidden opers.
* third/hiddenoper-tag implements the letspiss.net/hiddenoper
capability and message tag, which is added to all messages sent by
hidden opers. Only opers receive messages with this tag.
* third/id-tag implements the letspiss.net/id capability and message
tag, which is added to all messages and contains the SID or UID of the
sender. Only opers receive messages with this tag.