Change set::limit-svscmds default to ulines
This is done both in the default config and in the source code, so any
server updating to the latest PissIRCd version automatically has this
change.  Ordinary servers being able to use SVSJOIN and SVSNICK is
undesirable on pissnet and would cause desyncs if used since not every
server will have limit-svscmds set to servers.